Added cast for cents summation

+++ currency/currency.go
@@ -9,6 +9,6 @@ type Currency struct {
 }
 
 func (curr Currency) add(c1 Currency) {
-	curr.Dollars += c1.Dollars + (curr.Cents + c1.Cents) /  100
+	curr.Dollars += c1.Dollars + int((curr.Cents + c1.Cents) /  100)
 	curr.Cents += (curr.Cents + c1.Cents) % 100
 }
